test(e2e): build the e2e schema from migrations, not sync()
The e2e stack is the only suite that drives the real browser through
widget -> api -> console, and it was seeding a schema built from the
models via sync({ force: true }) — so the models were the source of truth
for both the code under test and the schema it ran against, and migration
drift stayed invisible to it.

Extracts the runner added in #38 to api/src/db/migrator.ts (e2e already
imports from api/src) and points both consumers at it: the API
integration harness and e2e/support/seed.ts.

This matters more than the integration change: the check job runs no
MySQL, so every integration test returns early on a null harness and
passes without asserting anything in CI. The e2e job brings up real
infra, so this is what actually closes the gap in CI.

Verified by removing the deleted_at migration: the journeys fail with
SequelizeDatabaseError: Unknown column 'deleted_at' at
POST /api/v1/visitor/chats — the same request that 500'd in the browser.
Restored: 8/8 journeys and 22/22 api tests pass.
